fix auto-reconnect by re-subscribing after connection established
